Project refactor/restructure
