+++ src/hooks/index.js
@@ -1,0 +1,2 @@
+export { default as useFetch } from './useFetch';
+export { default as useFetchUser } from './useFetchUser';
