+++ src/hooks/useFetch.js
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import { useEffect, useState } from 'react';
 
 const useFetch = (url, selectedCategory, selectedCountry, newEvent) => {
   const [data, setData] = useState(null);
